Change Elo property type in PlayerDto
Elo property type was changed from decimal to integer to simplify calculations and because the Elo rating system in our use case only needs integers. This change will help to reduce potential errors in the gameplay logic.

+++ Shared/Dtos/PlayerDto.cs
@@ -5,7 +5,7 @@ public class PlayerDto
     public Guid Id { get; set; }
     public string NickName { get; set; }
     public Uri? ImageUrl { get; set; }
-    public decimal Elo { get; set; }
+    public int Elo { get; set; }
     public bool CanIDeleteIt { get; set; }
     public bool DevelopmentAnnouncementReadIt { get; set; }
     public int GainedElo { get; set; }
